V 0.044
- Se agrego el editar cliente al ABMC, solo falta el borrar
- Se corrigio el error en el editar cliente.

+++ AgenciaDeAutos/Data/Interfaces/IDaoCliente.cs
@@ -14,6 +14,6 @@ namespace AgenciaDeAutos.Data.Interfaces
         bool nuevoCliente(C obj);
         bool editarCliente(C obj);
         bool borrarCliente(int id);
-        bool buscarClientePorDNI(int dni);
+        bool buscarClientePorDNI(long dni);
     }
 }
